add cpnversation reply, send message and moving files a bit
